+++ client/components/index.js
@@ -1,5 +1,5 @@
 /**
- * `components/index.js` exists simply as a 'central export' for our components.
+ * `components/index.js` exists as a 'central export' for our components.
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
@@ -11,6 +11,8 @@ export { default as TransactionReceipt } from './transactionReceipt'
 export { default as OpinionList } from './opinion-list'
 export { default as OpinionSelector } from './opinionSelector'
 export { default as CartForm } from './cartForm'
+export { default as CartFormTmpl } from './cartForm.tmpl'
+export { default as CartItem } from './cartItem.tmpl'
 export { default as Payment } from './payment'
 export { default as SimpleCard } from './card'
 export { default as ListGenerator } from './listGenerator'
